+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Added notice in developer console when C# evaluation is not supported.
+- Added command to change stats font size.
+- Changed warning text colour.
+- Changed stats_list output to be prettier.
+- Changed naming of stats created via the custom attribute.
+- Fixed exception thrown by using Mono on unsupported platforms.
 
 ## [1.0.2] - 2021-10-30
 - Added user-defined developer console stats which are displayed on-screen.